Interpolation

Interpolation

Types

DiffEqBase.AbstractDiffEqInterpolation — Type.

abstract type AbstractDiffEqInterpolation <: Function

DiffEqBase.ConstantInterpolation — Type.

struct ConstantInterpolation{T1, T2} <: DiffEqBase.AbstractDiffEqInterpolation

DiffEqBase.LinearInterpolation — Type.

struct LinearInterpolation{T1, T2} <: DiffEqBase.AbstractDiffEqInterpolation

DiffEqBase.HermiteInterpolation — Type.

struct HermiteInterpolation{T1, T2, T3} <: DiffEqBase.AbstractDiffEqInterpolation

Functions

DiffEqBase.interpolation — Function.

interpolation(tval, id, idxs, deriv, p)
interpolation(tval, id, idxs, deriv, p, continuity)

Get the value at tval where the solution is known at the times t (sorted), with values u and derivatives ks

DiffEqBase.interpolation! — Function.

interpolation!(vals, tvals, id, idxs, deriv, p)
interpolation!(vals, tvals, id, idxs, deriv, p, continuity)

Get the value at tvals where the solution is known at the times t (sorted), with values u and derivatives ks

interpolation!(out, tval, id, idxs, deriv, p)
interpolation!(out, tval, id, idxs, deriv, p, continuity)

Get the value at tval where the solution is known at the times t (sorted), with values u and derivatives ks

DiffEqBase.interpolant — Function.

Hairer Norsett Wanner Solving Ordinary Differential Equations I - Nonstiff Problems Page 190

Hermite Interpolation

Hermite Interpolation

Hermite Interpolation

Hermite Interpolation

Linear Interpolation

Constant Interpolation