Problems

Problems

Base types

DiffEqBase.DEProblem — Type.

abstract type DEProblem

Base type for all DifferentialEquations.jl problems. Concrete subtypes of DEProblem contain the necessary information to fully define a differential equation of the corresponding type.

Interface

DiffEqBase.remake — Function.

remake(thing; <keyword arguments>)

Re-construct thing with new field values specified by the keyword arguments.

Other functions

Missing docstring.

Missing docstring for DiffEqBase.get_concrete_problem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.get_concrete_u0. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.get_concrete_tspan. Check Documenter's build log for details.

DiffEqBase.promote_tspan — Function.

promote_tspan(tspan)

Convert the tspan field of a DEProblem to a (tmin, tmax) tuple, where both elements are of the same type. If tspan is a function, returns it as-is.

Missing docstring.

Missing docstring for DiffEqBase.eval_u0. Check Documenter's build log for details.

DiffEqBase.adaptive_warn — Function.

adaptive_warn(u0, tspan)

Check whether the values of u0 and tspan are appropriate for use with adaptive integrators and emit specific warnings if they are not.

DiffEqBase.adaptive_integer_warn — Function.

adaptive_integer_warn(tspan)

Emit a warning about incompatibility with adaptive integers if tspan contains integers.