Utility

Utility

Functions

DiffEqBase.num_types_in_tuple — Function.

num_types_in_tuple(sig)

Get the number of parameters of a Tuple type, i.e. the number of fields.

DiffEqBase.numargs — Function.

numargs(f)

Returns the number of arguments of f for the method which has the most arguments.

DiffEqBase.copy_fields — Function.

copy_fields(arr:AbstractArray, template::DEDataArray)

Create DEDataArray that wraps arr with all other fields set to a deep copy of the value in template.

copy_fields!(dest::T, src::T2) where {T<:DEDataArray,T2<:DEDataArray}

Replace all fields of dest except of its wrapped array with a copy of the value in src. Arrays are recursively copied.

DiffEqBase.undefined_exports — Function.

undefined_exports(mod)

List symbols export'ed but not actually defined.

DiffEqBase.warn_compat — Function.

warn_compat()

Emit a warning with a link to the solver compatibility chart in the documenation.

DiffEqBase.@def — Macro.

@def name definition

DiffEqBase.@add_kwonly — Macro.

@add_kwonly function_definition

Define keyword-only version of the function_definition.

@add_kwonly function f(x; y=1)
    ...
end

expands to:

function f(x; y=1)
    ...
end
function f(; x = error("No argument x"), y=1)
    ...
end

DiffEqBase.@CSI_str — Macro.

@CSI_str cmd

Create an ANSI escape sequence string for the CSI command cmd.