Solver Helpers

This package includes the documentation for the helper functions for the various solvers.

Abstract Types

DiffEqBase.DEIntegrator — Type.

DEIntegrator: A DifferentialEquations Integrator type, used to initiate a solver.

source

ODE Tableaus

Tableau Methods

Base.length — Method.

Base.length(tab::ODERKTableau)

Defines the length of a Runge-Kutta method to be the number of stages.

source

DiffEqDevTools.stability_region — Function.

stability_region(z,tab::ODERKTableau)

Calculates the stability function from the tableau at z. Stable if <1.

\[r(z) = \frac{\det(I-zA+zeb^T)}{\det(I-zA)}\]

source

DiffEqBase.ExplicitRKTableau — Type.

ExplicitRKTableau

Holds a tableau which defines an explicit Runge-Kutta method.

source

DiffEqBase.ImplicitRKTableau — Type.

ImplicitRKTableau

Holds a tableau which defines an implicit Runge-Kutta method.

source

DiffEqBase.ODERKTableau — Type.

ODERKTableau: A Runge-Kutta Tableau for an ODE integrator

source

OrdinaryDiffEq.ODE_DEFAULT_TABLEAU — Constant.

ODE_DEFAULT_TABLEAU

Sets the default tableau for the ODE solver. Currently Dormand-Prince 4/5.

source

Explicit Tableaus

DiffEqDevTools.constructEuler — Function.

Euler's method.

source

DiffEqDevTools.constructRalston — Function.

Ralston's Order 2 method.

source

DiffEqDevTools.constructHeun — Function.

Heun's Order 2 method.

source

DiffEqDevTools.constructKutta3 — Function.

Kutta's Order 3 method.

source

OrdinaryDiffEq.constructBS3 — Function.

constructBogakiShampine3()

Constructs the tableau object for the Bogakai-Shampine Order 2/3 method.

source

DiffEqDevTools.constructBogakiShampine3 — Function.

constructBogakiShampine3()

Constructs the tableau object for the Bogakai-Shampine Order 2/3 method.

source

DiffEqDevTools.constructRK4 — Function.

Classic RK4 method.

source

DiffEqDevTools.constructRK438Rule — Function.

Classic RK4 3/8's rule method.

source

DiffEqDevTools.constructRKF4 — Function.

Runge-Kutta-Fehberg Order 4/3

source

DiffEqDevTools.constructRKF5 — Function.

Runge-Kutta-Fehlberg Order 4/5 method.

source

DiffEqDevTools.constructCashKarp — Function.

constructCashKarp()

Constructs the tableau object for the Cash-Karp Order 4/5 method.

source

OrdinaryDiffEq.constructDormandPrince — Function.

constructDormandPrince()

Constructs the tableau object for the Dormand-Prince Order 4/5 method.

source

OrdinaryDiffEq.constructBS5 — Function.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

source

DiffEqDevTools.constructPapakostasPapaGeorgiou5 — Function.

S.N. Papakostas and G. PapaGeorgiou higher error more stable

A Family of Fifth-order Runge-Kutta Pairs, by S.N. Papakostas and G. PapaGeorgiou, Mathematics of Computation,Volume 65, Number 215, July 1996, Pages 1165-1181.

source

DiffEqDevTools.constructPapakostasPapaGeorgiou52 — Function.

S.N. Papakostas and G. PapaGeorgiou less stable lower error Strictly better than DP5

A Family of Fifth-order Runge-Kutta Pairs, by S.N. Papakostas and G. PapaGeorgiou, Mathematics of Computation,Volume 65, Number 215, July 1996, Pages 1165-1181.

source

DiffEqDevTools.constructTsitouras5 — Function.

Runge–Kutta pairs of orders 5(4) using the minimal set of simplifying assumptions, by Ch. Tsitouras, TEI of Chalkis, Dept. of Applied Sciences, GR34400, Psahna, Greece.

source

DiffEqDevTools.constructLutherKonen5 — Function.

Luther and Konen's First Order 5 Some Fifth-Order Classical Runge Kutta Formulas, H.A.Luther and H.P.Konen, Siam Review, Vol. 3, No. 7, (Oct., 1965) pages 551-558.

source

DiffEqDevTools.constructLutherKonen52 — Function.

Luther and Konen's Second Order 5 Some Fifth-Order Classical Runge Kutta Formulas, H.A.Luther and H.P.Konen, Siam Review, Vol. 3, No. 7, (Oct., 1965) pages 551-558.

source

DiffEqDevTools.constructLutherKonen53 — Function.

Luther and Konen's Third Order 5 Some Fifth-Order Classical Runge Kutta Formulas, H.A.Luther and H.P.Konen, Siam Review, Vol. 3, No. 7, (Oct., 1965) pages 551-558.

source

DiffEqDevTools.constructRungeFirst5 — Function.

Runge's First Order 5 method

source

DiffEqDevTools.constructLawson5 — Function.

Lawson's 5th order scheme

An Order Five Runge Kutta Process with Extended Region of Stability, J. Douglas Lawson, Siam Journal on Numerical Analysis, Vol. 3, No. 4, (Dec., 1966) pages 593-597

source

DiffEqDevTools.constructSharpSmart5 — Function.

Explicit Runge-Kutta Pairs with One More Derivative Evaluation than the Minimum, by P.W.Sharp and E.Smart, Siam Journal of Scientific Computing, Vol. 14, No. 2, pages. 338-348, March 1993.

source

DiffEqDevTools.constructBogakiShampine5 — Function.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

source

DiffEqDevTools.constructCassity5 — Function.

Cassity's Order 5 method

source

DiffEqDevTools.constructButcher6 — Function.

Butcher's First Order 6 method

On Runge-Kutta Processes of High Order, by J. C. Butcher, Journal of the Australian Mathematical Society, Vol. 4, (1964), pages 179 to 194

source

DiffEqDevTools.constructButcher62 — Function.

Butcher's Second Order 6 method

On Runge-Kutta Processes of High Order, by J. C. Butcher, Journal of the Australian Mathematical Society, Vol. 4, (1964), pages 179 to 194

source

DiffEqDevTools.constructButcher63 — Function.

Butcher's Third Order 6

On Runge-Kutta Processes of High Order, by J. C. Butcher, Journal of the Australian Mathematical Society, Vol. 4, (1964), pages 179 to 194

source

DiffEqDevTools.constructVernerRobust6 — Function.

From Verner's Website

source

DiffEqDevTools.constructTanakaKasugaYamashitaYazaki6A — Function.

TanakaKasugaYamashitaYazaki Order 6 A

On the Optimization of Some Eight-stage Sixth-order Explicit Runge-Kutta Method, by M. Tanaka, K. Kasuga, S. Yamashita and H. Yazaki, Journal of the Information Processing Society of Japan, Vol. 34, No. 1 (1993), pages 62 to 74.

source

DiffEqDevTools.constructTanakaKasugaYamashitaYazaki6B — Function.

constructTanakaKasugaYamashitaYazaki Order 6 B

On the Optimization of Some Eight-stage Sixth-order Explicit Runge-Kutta Method, by M. Tanaka, K. Kasuga, S. Yamashita and H. Yazaki, Journal of the Information Processing Society of Japan, Vol. 34, No. 1 (1993), pages 62 to 74.

source

DiffEqDevTools.constructTanakaKasugaYamashitaYazaki6C — Function.

constructTanakaKasugaYamashitaYazaki Order 6 C

On the Optimization of Some Eight-stage Sixth-order Explicit Runge-Kutta Method, by M. Tanaka, K. Kasuga, S. Yamashita and H. Yazaki, Journal of the Information Processing Society of Japan, Vol. 34, No. 1 (1993), pages 62 to 74.

source

DiffEqDevTools.constructTanakaKasugaYamashitaYazaki6D — Function.

constructTanakaKasugaYamashitaYazaki Order 6 D

On the Optimization of Some Eight-stage Sixth-order Explicit Runge-Kutta Method, by M. Tanaka, K. Kasuga, S. Yamashita and H. Yazaki, Journal of the Information Processing Society of Japan, Vol. 34, No. 1 (1993), pages 62 to 74.

source

DiffEqDevTools.constructHuta6 — Function.

Anton Hutas First Order 6 method

Une amélioration de la méthode de Runge-Kutta-Nyström pour la résolution numérique des équations différentielles du premièr ordre, by Anton Huta, Acta Fac. Nat. Univ. Comenian Math., Vol. 1, pages 201-224 (1956).

source

DiffEqDevTools.constructHuta62 — Function.

Anton Hutas Second Order 6 method

Une amélioration de la méthode de Runge-Kutta-Nyström pour la résolution numérique des équations différentielles du premièr ordre, by Anton Huta, Acta Fac. Nat. Univ. Comenian Math., Vol. 1, pages 201-224 (1956).

source

DiffEqDevTools.constructVerner6 — Function.

Verner Order 5/6 method

A Contrast of a New RK56 pair with DP56, by Jim Verner, Department of Mathematics. Simon Fraser University, Burnaby, Canada, 2006.

source

DiffEqDevTools.constructDormandPrince6 — Function.

Dormand-Prince Order 5//6 method

P.J. Prince and J. R. Dormand, High order embedded Runge-Kutta formulae, Journal of Computational and Applied Mathematics . 7 (1981), pp. 67-75.

source

DiffEqDevTools.constructSharpVerner6 — Function.

Sharp-Verner Order 5/6 method

Completely Imbedded Runge-Kutta Pairs, by P. W. Sharp and J. H. Verner, SIAM Journal on Numerical Analysis, Vol. 31, No. 4. (Aug., 1994), pages. 1169 to 1190.

source

OrdinaryDiffEq.constructVern6 — Function.

From Verner's Website

source

DiffEqDevTools.constructClassicVerner6 — Function.

EXPLICIT RUNGE-KUTFA METHODS WITH ESTIMATES OF THE LOCAL TRUNCATION ERROR

source

DiffEqDevTools.constructChummund6 — Function.

Chummund's First Order 6 method

A three-dimensional family of seven-step Runge-Kutta methods of order 6, by G. M. Chammud (Hammud), Numerical Methods and programming, 2001, Vol.2, 2001, pages 159-166 (Advanced Computing Scientific journal published by the Research Computing Center of the Lomonosov Moscow State Univeristy)

source

DiffEqDevTools.constructChummund62 — Function.

Chummund's Second Order 6 method

A three-dimensional family of seven-step Runge-Kutta methods of order 6, by G. M. Chammud (Hammud), Numerical Methods and programming, 2001, Vol.2, 2001, pages 159-166 (Advanced Computing Scientific journal published by the Research Computing Center of the Lomonosov Moscow State Univeristy)

source

DiffEqDevTools.constructPapakostas6 — Function.

Papakostas's Order 6

On Phase-Fitted modified Runge-Kutta Pairs of order 6(5), by Ch. Tsitouras and I. Th. Famelis, International Conference of Numerical Analysis and Applied Mathematics, Crete, (2006)

source

DiffEqDevTools.constructLawson6 — Function.

Lawson's Order 6

An Order 6 Runge-Kutta Process with an Extended Region of Stability, by J. D. Lawson, Siam Journal on Numerical Analysis, Vol. 4, No. 4 (Dec. 1967) pages 620-625.

source

DiffEqDevTools.constructTsitourasPapakostas6 — Function.

Tsitouras-Papakostas's Order 6

Cheap Error Estimation for Runge-Kutta methods, by Ch. Tsitouras and S.N. Papakostas, Siam Journal on Scientific Computing, Vol. 20, Issue 6, Nov 1999.

source

DiffEqDevTools.constructDormandLockyerMcCorriganPrince6 — Function.

DormandLockyerMcCorriganPrince Order 6 Global Error Estimation

Global Error estimation with Runge-Kutta triples, by J.R.Dormand, M.A.Lockyer, N.E.McCorrigan and P.J.Prince, Computers and Mathematics with Applications, 18 (1989) pages 835-846.

source

DiffEqDevTools.constructVernerEfficient6 — Function.

From Verner's Website

source

DiffEqDevTools.constructMikkawyEisa — Function.

Mikkawy-Eisa Order 6

A general four-parameter non-FSAL embedded Runge–Kutta algorithm of orders 6 and 4 in seven stages, by M.E.A. El-Mikkawy and M.M.M. Eisa, Applied Mathematics and Computation, Vol. 143, No. 2, (2003) pages 259 to 267.

source

DiffEqDevTools.constructVernerEfficient7 — Function.

From Verner's website

source

DiffEqDevTools.constructClassicVerner7 — Function.

EXPLICIT RUNGE-KUTFA METHODS WITH ESTIMATES OF THE LOCAL TRUNCATION ERROR

source

DiffEqDevTools.constructSharpVerner7 — Function.

Completely Imbedded Runge-Kutta Pairs, by P.W.Sharp and J.H.Verner, Siam Journal on Numerical Analysis, Vol.31, No.4. (August 1994) pages 1169-1190.

source

DiffEqDevTools.constructTanakaYamashitaStable7 — Function.

On the Optimization of Some Nine-Stage Seventh-order Runge-Kutta Method, by M. Tanaka, S. Muramatsu and S. Yamashita, Information Processing Society of Japan, Vol. 33, No. 12 (1992) pages 1512-1526.

source

DiffEqDevTools.constructSharpSmart7 — Function.

Explicit Runge-Kutta Pairs with One More Derivative Evaluation than the Minimum, by P.W.Sharp and E.Smart, Siam Journal of Scientific Computing, Vol. 14, No. 2, pages. 338-348, March 1993.

source

DiffEqDevTools.constructTanakaYamashitaEfficient7 — Function.

On the Optimization of Some Nine-Stage Seventh-order Runge-Kutta Method, by M. Tanaka, S. Muramatsu and S. Yamashita, Information Processing Society of Japan, Vol. 33, No. 12 (1992) pages 1512-1526.

source

DiffEqDevTools.constructVernerRobust7 — Function.

From Verner's website

source

OrdinaryDiffEq.constructTanYam7 — Function.

On the Optimization of Some Nine-Stage Seventh-order Runge-Kutta Method, by M. Tanaka, S. Muramatsu and S. Yamashita, Information Processing Society of Japan, Vol. 33, No. 12 (1992) pages 1512-1526.

source

DiffEqDevTools.constructEnrightVerner7 — Function.

The Relative Efficiency of Alternative Defect Control Schemes for High-Order Continuous Runge-Kutta Formulas W. H. Enright SIAM Journal on Numerical Analysis, Vol. 30, No. 5. (Oct., 1993), pp. 1419-1445.

source

DiffEqDevTools.constructDormandPrince8 — Function.

constructDormandPrice8()

Constructs the tableau object for the Dormand-Prince Order 6/8 method.

source

DiffEqDevTools.constructRKF8 — Function.

constructRKF8()

Constructs the tableau object for the Runge-Kutta-Fehlberg Order 7/8 method.

source

DiffEqDevTools.constructCooperVerner8 — Function.

Some Explicit Runge-Kutta Methods of High Order, by G. J. Cooper and J. H. Verner, SIAM Journal on Numerical Analysis, Vol. 9, No. 3, (September 1972), pages 389 to 405

source

DiffEqDevTools.constructCooperVerner82 — Function.

Some Explicit Runge-Kutta Methods of High Order, by G. J. Cooper and J. H. Verner, SIAM Journal on Numerical Analysis, Vol. 9, No. 3, (September 1972), pages 389 to 405

source

DiffEqDevTools.constructTsitourasPapakostas8 — Function.

Cheap Error Estimation for Runge-Kutta methods, by Ch. Tsitouras and S.N. Papakostas, Siam Journal on Scientific Computing, Vol. 20, Issue 6, Nov 1999.

source

DiffEqDevTools.constructEnrightVerner8 — Function.

The Relative Efficiency of Alternative Defect Control Schemes for High-Order Continuous Runge-Kutta Formulas W. H. Enright SIAM Journal on Numerical Analysis, Vol. 30, No. 5. (Oct., 1993), pp. 1419-1445.

source

DiffEqDevTools.constructdverk78 — Function.

Jim Verner's "Maple" (dverk78)

source

DiffEqDevTools.constructClassicVerner8 — Function.

EXPLICIT RUNGE-KUTFA METHODS WITH ESTIMATES OF THE LOCAL TRUNCATION ERROR

source

DiffEqDevTools.constructDormandPrince8_64bit — Function.

constructDormandPrice8_64bit()

Constructs the tableau object for the Dormand-Prince Order 6/8 method with the approximated coefficients from the paper. This works until below 64-bit precision.

source

DiffEqDevTools.constructCurtis8 — Function.

An Eighth Order Runge-Kutta process with Eleven Function Evaluations per Step, by A. R. Curtis, Numerische Mathematik, Vol. 16, No. 3 (1970), pages 268 to 277

source

OrdinaryDiffEq.constructTsitPap8 — Function.

Cheap Error Estimation for Runge-Kutta methods, by Ch. Tsitouras and S.N. Papakostas, Siam Journal on Scientific Computing, Vol. 20, Issue 6, Nov 1999.

source

DiffEqDevTools.constructSharp9 — Function.

Journal of Applied Mathematics & Decision Sciences, 4(2), 183-192 (2000), "High order explicit Runge-Kutta pairs for ephemerides of the Solar System and the Moon".

source

DiffEqDevTools.constructTsitouras9 — Function.

Optimized explicit Runge-Kutta pairs of order 9(8), by Ch. Tsitouras, Applied Numerical Mathematics, 38 (2001) 123-134.

source

DiffEqDevTools.constructTsitouras92 — Function.

Optimized explicit Runge-Kutta pairs of order 9(8), by Ch. Tsitouras, Applied Numerical Mathematics, 38 (2001) 123-134.

source

DiffEqDevTools.constructVernerEfficient9 — Function.

From Verner's Webiste

source

OrdinaryDiffEq.constructVern9 — Function.

From Verner's Webiste

source

DiffEqDevTools.constructVerner916 — Function.

Verner 1991 First Order 5/6 method

Some Ruge-Kutta Formula Pairs, by J.H.Verner, SIAM Journal on Numerical Analysis, Vol. 28, No. 2 (April 1991), pages 496 to 511.

source

DiffEqDevTools.constructVerner9162 — Function.

Verner 1991 Second Order 5/6 method

Some Ruge-Kutta Formula Pairs, by J.H.Verner, SIAM Journal on Numerical Analysis, Vol. 28, No. 2 (April 1991), pages 496 to 511.

source

DiffEqDevTools.constructVernerRobust9 — Function.

From Verner's Webiste

source

OrdinaryDiffEq.constructFeagin10 — Function.

constructFeagin10

source

DiffEqDevTools.constructFeagin10Tableau — Function.

Feagin10 in Tableau form

source

DiffEqDevTools.constructOno10 — Function.

Ono10

source

DiffEqDevTools.constructCurtis10 — Function.

High-order Explicit Runge-Kutta Formulae, Their uses, and Limitations, A.R.Curtis, J. Inst. Maths Applics (1975) 16, 35-55.

source

DiffEqDevTools.constructHairer10 — Function.

A Runge-Kutta Method of Order 10, E. Hairer, J. Inst. Maths Applics (1978) 21, 47-59.

source

DiffEqDevTools.constructBaker10 — Function.

Tom Baker, University of Teeside. Part of RK-Aid http://www.scm.tees.ac.uk/users/u0000251/research/researcht.htm http://www.scm.tees.ac.uk/users/u0000251/j.r.dormand/t.baker/rk10921m/rk10921m

source

OrdinaryDiffEq.constructFeagin12 — Function.

constructFeagin12

source

DiffEqDevTools.constructOno12 — Function.

On the 25 stage 12th order explicit Runge-Kutta method, by Hiroshi Ono. Transactions of the Japan Society for Industrial and applied Mathematics, Vol. 6, No. 3, (2006) pages 177 to 186

source

DiffEqDevTools.constructFeagin12Tableau — Function.

Tableau form of Feagin12

source

OrdinaryDiffEq.constructFeagin14 — Function.

constructFeagin14

source

DiffEqDevTools.constructFeagin14Tableau — Function.

Tableau form of Feagin14

source

Implicit Tableaus

DiffEqDevTools.constructImplicitEuler — Function.

Implicit Euler Method

source

DiffEqDevTools.constructMidpointRule — Function.

Order 2 Midpoint Method

source

DiffEqDevTools.constructTrapezoidalRule — Function.

Order 2 Trapezoidal Rule (LobattoIIIA2)

source

DiffEqDevTools.constructLobattoIIIA4 — Function.

LobattoIIIA Order 4 method

source

DiffEqDevTools.constructLobattoIIIB2 — Function.

LobattoIIIB Order 2 method

source

DiffEqDevTools.constructLobattoIIIB4 — Function.

LobattoIIIB Order 4 method

source

DiffEqDevTools.constructLobattoIIIC2 — Function.

LobattoIIIC Order 2 method

source

DiffEqDevTools.constructLobattoIIIC4 — Function.

LobattoIIIC Order 4 method

source

DiffEqDevTools.constructLobattoIIICStar2 — Function.

LobattoIIIC* Order 2 method

source

DiffEqDevTools.constructLobattoIIICStar4 — Function.

LobattoIIIC* Order 4 method

source

DiffEqDevTools.constructLobattoIIID2 — Function.

LobattoIIID Order 2 method

source

DiffEqDevTools.constructLobattoIIID4 — Function.

LobattoIIID Order 4 method

source

DiffEqDevTools.constructGL2 — Function.

Gauss-Legendre Order 2.

source

DiffEqDevTools.constructGL4 — Function.

Gauss-Legendre Order 4.

source

DiffEqDevTools.constructGL6 — Function.

Gauss-Legendre Order 6.

source

DiffEqDevTools.constructRadauIA3 — Function.

RadauIA Order 3 method

source

DiffEqDevTools.constructRadauIA5 — Function.

RadauIA Order 5 method

source

DiffEqDevTools.constructRadauIIA3 — Function.

RadauIIA Order 3 method

source

DiffEqDevTools.constructRadauIIA5 — Function.

RadauIIA Order 5 method

source

SDE Solver Extras

StochasticDiffEq.monteCarloSim — Function.

monteCarloSim(Δt::Number,prob::SDEProblem)

Performs a parallel Monte-Carlo simulation to solve the SDE problem with Δt numMonte times. Returns a vector of solution objects.

Keyword Arguments

T - Final time. Default is 1.

numMonte - Number of Monte-Carlo simulations to run. Default is 10000

save_timeseries - Denotes whether save_timeseries should be turned on in each run. Default is false.

source

StochasticDiffEq.RosslerSRI — Type.

RosslerSRI

Holds the Butcher tableaus for a Rosser SRI method.

source

StochasticDiffEq.RosslerSRA — Type.

RosslerSRA

Holds the Butcher tableaus for a Rosser SRA method.

source

StochasticDiffEq.constructSRA1 — Function.

constructSRA1()

Constructs the taleau type for the SRA1 method.

source

StochasticDiffEq.constructSRIW1 — Function.

constructSRIW1()

Constructs the tableau type for the SRIW1 method.

source

StochasticDiffEq.checkSRAOrder — Function.

checkSRAOrder(RosslerSRI)

Determines whether the order conditions are met via the tableaus of the SRA method.

source

StochasticDiffEq.checkSRIOrder — Function.

checkSRIOrder(RosslerSRI)

Determines whether the order conditions are met via the tableaus of the SRI method.

source

Stationary Stokes

StokesDiffEq.GSδq! — Function.

GSδq!(δq,rp,Δxs)

Performs a Gauss-Seidel iteration for δq.

source

StokesDiffEq.GSu! — Function.

GSu!(u,f₁,Δxs,p,ugD,grids,ux,uy)

Performs a Gauss-Seidel iteration on u.

source

StokesDiffEq.calc_rp! — Function.

calc_rp!(rp,u,v,Δxs,g,px,py)

Calculates the rp from the u and v's.

source

StokesDiffEq.update_p! — Function.

update_p!(p,δq,Δxs)

Updates p given δq

source

StokesDiffEq.update_v! — Function.

update_v!(v,δq,Δxs)

Updates v given δq

source

StokesDiffEq.uzawa_p! — Function.

uzawa_p!(p,u,v,Δxs,g,px,py)

Solves for p from u and v using an Uzawa update.

source

StokesDiffEq.stokes_restriction — Function.

stokes_restriction(u,v,p,Δxs,grids,mins,maxs,ugD,vgD)

Restricts the Stokes problem to the coarsegrid.

source

StokesDiffEq.stokes_prolongation — Function.

stokes_prolongation(u,v,p,Δxs,grids,mins,maxs,ugD,vgD)

Prolongates the Stokes problem to the fine grid

source

StokesDiffEq.update_u! — Function.

update_u!(u,δq,Δxs)

Updates u given δq

source

StokesDiffEq.GSv! — Function.

GSv!(v,f₂,Δxs,p,vgD,grids,vx,vy)

Performs a Gauss-Seidel iteration on v.

source