Defining a SDE Problem

To define an SDE Problem, you simply need to give the forcing function $f$, the noise function g, and the initial condition $u₀$ which define an SDE

\[du = f(t,u)dt + Σgᵢ(t,u)dWⁱ\]

f and g should be specified as f(t,u) and g(t,u) respectively, and u₀ should be an AbstractArray whose geometry matches the desired geometry of u. Note that we are not limited to numbers or vectors for u₀, one is allowed to provide u₀ as arbitrary matrices / higher dimension tensors as well. A vector of gs can also be defined to determine an SDE of higher Ito dimension.

Problem Type

StochasticDiffEq.SDEProblem — Type.

SDEProblem

Wraps the data which defines an SDE problem

\[u = f(u,t)dt + Σgᵢ(u,t)dWⁱ\]

with initial condition $u₀$.

Constructors

SDEProblem(f,g,u₀;analytic=nothing) : Defines the SDE with the specified functions and defines the solution if analytic is given.

Fields

f: The drift function in the SDE.

g: The noise function in the SDE.

u₀: The initial condition.

analytic: A function which describes the solution.

knownanalytic: True if the solution is given.

numvars: The number of variables in the system

sizeu: The size of the initial condition (and thus u)

noise: The noise process applied to the noise upon generation.

source

Example Problems

Examples problems can be found in src/premades/premade_problems.jl

DiffEqProblemLibrary.prob_sde_linear — Constant.

\[du_t = βudt + αudW_t\]

where β=1.01, α=0.87, and initial condtion u₀=1/2, with solution

\[u(t,u₀,W_t)=u₀\exp((α-\frac{β^2}{2})t+βW_t)\]

source

DiffEqProblemLibrary.prob_sde_2Dlinear — Constant.

8 linear SDEs (as a 4x2 matrix):

\[du_t = βudt + αudW_t\]

where β=1.01, α=0.87, and initial condtion u₀=1/2 with solution

\[u(t,u₀,W_t)=u₀\exp((α-\frac{β^2}{2})t+βW_t)\]

source

DiffEqProblemLibrary.prob_sde_wave — Constant.

\[du_t = -\frac{1}{100}sin(u)cos^3(u)dt + \frac{1}{10}cos^{2}(u_t) dW_t\]

and initial condition u₀=1.0 with solution

\[u(t,u₀,W_t)=\arctan(\frac{W_t}{10} + \tan(u₀))\]

source

DiffEqProblemLibrary.prob_sde_lorenz — Constant.

Lorenz Attractor with additive noise

\[\begin{align} dx &= σ*(y-x)dt + αdW_t \\ dy &= (x*(ρ-z) - y)dt + αdW_t \\ dz &= (x*y - β*z)dt + αdW_t \\ \end{align}\]

with $σ=10$, $ρ=28$, $β=8/3$, $α=3.0$ and inital condition $u₀=[1;1;1]$.

source

DiffEqProblemLibrary.prob_sde_cubic — Constant.

\[du_t = \frac{1}{4}u(1-u^2)dt + \frac{1}{2}(1-u^2)dW_t\]

and initial condtion u₀=1/2, with solution

\[u(t,u₀,W_t)=\frac{(1+u₀)\exp(W_t)+u₀-1}{(1+u₀)\exp(W_t)+1-u₀}\]

source

DiffEqProblemLibrary.prob_sde_additive — Constant.

Additive noise problem

\[u_t = (\frac{β}{\sqrt{1+t}}-\frac{1}{2(1+t)}u_t)dt + \frac{αβ}{\sqrt{1+t}}dW_t\]

and initial condition u₀=1.0 with α=0.1 and β=0.05, with solution

\[u(t,u₀,W_t)=\frac{u₀}{\sqrt{1+t}} + \frac{β(t+αW_t)}{\sqrt{1+t}}\]

source

DiffEqProblemLibrary.prob_sde_additivesystem — Constant.

A multiple dimension extension of additiveSDEExample

source