Defining a Stokes Problem

Problem Type

StokesDiffEq.StokesProblem — Type.

StokesProblem Defines the solution to a stationary Stokes problem:

\[\]

Constructors

StokesProblem(f₁,f₂,g,uanalytic,vanalytic,panalytic) StokesProblem(f₁,f₂,g,ugD,vgD)

Fields

f₁::Function

f₂::Function

g::Function

ugD::Function

vgD::Function

uanalytic::Function

vanalytic::Function

panalytic::Function

trueknown::Bool

source

Example Problems

Examples problems can be found in src/premades/premade_problems.jl

StokesDiffEq.prob_stokes_dirichletzero — Constant.

Example problem for solving the trivial stationary Stokes equation.

source

StokesDiffEq.prob_stokes_homogenous — Constant.

Example problem for a homogeneous stationary Stokes equation.

source