The Solution Type

Each solver has an appropriate solution type. The solution type holds all of the information about the problem which was solved an its solution. If you enabled save_timeseries=true, then the solver also includes a time-course of the solution captured at every timeseries_steps steps.

The solution type has a lot of built in functionality to help analysis. For example, it has an array interface for accessing the values. We can use

sol[i]

to access the value at timestep i (if the timeseres was saved), and

sol.t[i]

to access the value of t at timestep i.

If the solver allows for dense output (any ODE solver) and dense=true was set for the solving (which is the default), then we can access the approximate value at a time t using the command

sol(t)

If the analytical solution, we also have

sol.u_analytic # timeseries of analytical solution
sol.prob.analytic(t) # The analytic solution at time t

Plotting functionality is provided for each solution type. To plot the solution, simply use

plot(sol)

The plotting function is implemented as a recipe to Plots.jl and as such receives all of the features of a Plots.jl plot.

Related Functions

DiffEqDevTools.appxtrue! — Function.

appxtrue!(sol::FEMSolution,sol2::FEMSolution)

Adds the solution from sol2 to the FEMSolution object sol. Useful to add a quasi-true solution when none is known by computing once at a very small time/space step and taking that solution as the "true" solution

source

FiniteElementDiffEq.FEMSolutionTS — Function.

S = FEMSolutionTS(timeseries::Vector{uType},numvars::Int)S[i][j]` => Variable i at time j.

source