Finite Element Method Heat Equation Solvers

Avaliable Methods

[method] denotes an additional version for handling stochastic partial differential equations.

Finite Element Solvers (Stochastic) PDEs

Semilinear Heat Equation (Reaction-Diffusion)

Forward Euler [Maruyama]

Backward Euler [Maruyama]

Semi-implicit Crank-Nicholson [Maruyama]

Semi-implicit Backward Euler [Maruyama]

Linear Heat Equation

Forward Euler [Maruyama]

Backward Euler [Maruyama]

Crank-Nicholson [Maruyama]

Implicit Solvers

Direct

Factorizations (LU, Cholesky, QR, SVD)

Conjugate-Gradient (CG)

GMRES

Solver Documentation

DiffEqBase.solve — Method.

Finite Element Heat Equation Solver

solve(fem_mesh::FEMmesh,pdeProb::HeatProblem)

Takes in a definition for the heat equation $u_t = Δu + f$ on fem_mesh with functions as defined in pdeProb. If σ is specified in pdeProb, then this solves the stochastic heat equation $u_t = Δu + f + σdW_t$.

Keyword Arguments

alg = Solution algorithm. Default is :Euler. The choices are:

Linear

:Euler (Explicit)

:ImplicitEuler (Implicit)

:CrankNicholson (Implicit)

Nonlinear

:Euler (Explicit)

:ImplicitEuler (Nonlinear Solve)

:CrankNicholson (Nonlinear Solve)

:SemiImplicitEuler (Implicit)

:SemiImplicitCrankNicholson (Implicit)

Explicit algorithms only require solving matrix multiplications Au. Implicit algorithms require solving the linear equation Ax=b where x is the unknown. Nonlinear Solve algorithms require solving the nonlinear equation f(x)=0 using methods like Newton's method and is provided by NLSolve.jl. Explicit algorithms have the least stability and should be used either small Δt and non-stiff equations. The implicit algorithms have better stability, but for nonlinear equations require costly nonlinear solves in order to be solved exactly. The semi-implicit algorithms discretize with part of the equation implicit and another part explicit in order to allow for the algorithm to not require a nonlinear solve, but at the cost of some stability (though still vastly better at stability than explicit algorithms).

solver = Linear solver algorithm. This is the algorithm which is chosen for solving the implicit equation Ax=b. The default is LU. The choices are:

:Direct = Solves using \ (no factorization). Not recommended.

:Cholesky = Cholsky decomposition. Only stable of $I ± ΔtM⁻¹A$ is positive definite. This means that this works best when Δt is small. When applicable, this is the fastest.

:LU = LU-Decomposition. A good mix between fast and stable.

:QR = QR-Decomposition. Less numerical roundoff error than LU, but slightly slower.

:SVD = SVD-Decomposition. By far the slowest, but the most robust to roundoff error.

:CG = Conjugate-Gradient. Best when the space is very large and $I ± ΔtM⁻¹A$ is positive definite.

:GMRES = GMRES. Best when the space is very large and $I ± ΔtM⁻¹A$ is not positive definite.

save_timeseries = Makes the algorithm save the output at every timeseries_steps timesteps. By default save_timeseries is false.

timeseries_steps = If save_timeseries=true, then this is the number of steps between the saves.

autodiff = Whether or not autodifferentiation (as provided by AutoDiff.jl) is used for the nonlinear solving. By default autodiff is false.

method = Method the nonlinear solver uses. Defaults to :trust_region.

show_trace = Whether to show the output of the nonlinear solver. Defaults to false.

iterations = Maximum numer of iterations in the nonlinear solver. Defaults to 1000.

progress_steps = The number of steps between updates of the progress bar. Defaults to 1000.

progressbar = Turns on/off use of the Juno progress bar. Defaults to true. Requires Juno.

source