DAE Problems

SciMLBase.DAEProblem — Type

Defines an implicit ordinary differential equation (ODE) or differential-algebraic equation (DAE) problem. Documentation Page: https://diffeq.sciml.ai/stable/types/dae_types/

Mathematical Specification of an DAE Problem

To define a DAE Problem, you simply need to give the function $f$ and the initial condition $u_0$ which define an ODE:

\[0 = f(du,u,p,t)\]

f should be specified as f(du,u,p,t) (or in-place as f(resid,du,u,p,t)). Note that we are not limited to numbers or vectors for u₀; one is allowed to provide u₀ as arbitrary matrices / higher dimension tensors as well.

Problem Type

Constructors

DAEProblem(f::DAEFunction,du0,u0,tspan,p=NullParameters();kwargs...)

DAEProblem{isinplace}(f,du0,u0,tspan,p=NullParameters();kwargs...) : Defines the DAE with the specified functions. isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Parameters are optional, and if not given then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

For specifying Jacobians and mass matrices, see the DiffEqFunctions page.

Fields

f: The function in the ODE.

du0: The initial condition for the derivative.

u0: The initial condition.

tspan: The timespan for the problem.

differential_vars: A logical array which declares which variables are the differential (non algebraic) vars (i.e. du' is in the equations for this variable). Defaults to nothing. Some solvers may require this be set if an initial condition needs to be determined.

p: The parameters for the problem. Defaults to NullParameters

kwargs: The keyword arguments passed onto the solves.

Example Problems

Examples problems can be found in DiffEqProblemLibrary.jl.

To use a sample problem, such as prob_dae_resrob, you can do something like:

#] add DiffEqProblemLibrary
using DiffEqProblemLibrary.DAEProblemLibrary
# load problems
DAEProblemLibrary.importdaeproblems()
prob = DAEProblemLibrary.prob_dae_resrob
sol = solve(prob,IDA())

Missing docstring.

Missing docstring for DAEFunction. Check Documenter's build log for details.

Solution Type

SciMLBase.DAESolution — Type

struct DAESolution{T, N, uType, duType, uType2, DType, tType, P, A, ID, DE} <: SciMLBase.AbstractDAESolution{T, N, uType}

Representation of the solution to an differential-algebraic equation defined by an DAEProblem.

DESolution Interface

For more information on interacting with DESolution types, check out the Solution Handling page of the DifferentialEquations.jl documentation.

https://diffeq.sciml.ai/stable/basics/solution/

Fields

u: the representation of the DAE solution. Given as an array of solutions, where u[i] corresponds to the solution at time t[i]. It is recommended in most cases one does not access sol.u directly and instead use the array interface described in the Solution Handling page of the DifferentialEquations.jl documentation.

du: the representation fo the derivatives of the DAE solution.

t: the time points corresponding to the saved values of the DAE solution.

prob: the original DAEProblem that was solved.

alg: the algorithm type used by the solver.

destats: statistics of the solver, such as the number of function evaluations required, number of Jacobians computed, and more.

retcode: the return code from the solver. Used to determine whether the solver solved successfully (sol.retcode === :Success), whether it terminated due to a user-defined callback (sol.retcode === :Terminated), or whether it exited due to an error. For more details, see the return code section of the DifferentialEquations.jl documentation.

Example Problems

Examples problems can be found in DiffEqProblemLibrary.jl.

To use a sample problem, such as prob_dae_resrob, you can do something like:

#] add DiffEqProblemLibrary
using DiffEqProblemLibrary.DAEProblemLibrary
# load problems
DAEProblemLibrary.importdaeproblems()
prob = DAEProblemLibrary.prob_dae_resrob
sol = solve(prob,IDA())

DiffEqProblemLibrary.DAEProblemLibrary.prob_dae_resrob — Constant

The Robertson biochemical reactions in DAE form

\[\frac{dy₁}{dt} = -k₁y₁+k₃y₂y₃\]

\[\frac{dy₂}{dt} = k₁y₁-k₂y₂^2-k₃y₂y₃\]

\[1 = y₁ + y₂ + y₃\]

where $k₁=0.04$, $k₂=3\times10^7$, $k₃=10^4$. For details, see: Hairer Norsett Wanner Solving Ordinary Differential Equations I - Nonstiff Problems Page 129 Usually solved on $[0,1e11]$