Integrators

Integrators

Types

DiffEqBase.DEIntegrator — Type.

abstract type DEIntegrator

DiffEqBase.IntegratorIntervals — Type.

struct IntegratorIntervals{I}

Missing docstring.

Missing docstring for DiffEqBase.IntegratorTuples. Check Documenter's build log for details.

Interface

Missing docstring.

Missing docstring for DiffEqBase.initialize!(u, t, integrator::DiffEqBase.DEIntegrator, any_modified, c). Check Documenter's build log for details.

DiffEqBase.step! — Function.

step!(integ::DEIntegrator [, dt [, stop_at_tdt]])

Perform one (successful) step on the integrator.

Alternative, if a dt is given, then step! the integrator until there is a temporal difference ≥ dt in integ.t. When true is passed to the optional third argument, the integrator advances exactly dt.

Missing docstring.

Missing docstring for DiffEqBase.addat!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.get_tmp_cache. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.user_cache. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.u_cache. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.du_cache. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.ratenoise_cache. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.rand_cache. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.full_cache. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.resize_non_user_cache!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.deleteat_non_user_cache!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.addat_non_user_cache!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.terminate!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.get_du. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.get_du!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.get_dt. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.get_proposed_dt. Check Documenter's build log for details.

DiffEqBase.u_modified! — Function.

savevalues!(integrator::DEIntegrator,
  force_save=false) -> Tuple{Bool, Bool}

Try to save the state and time variables at the current time point, or the saveat point by using interpolation when appropriate. It returns a tuple that is (saved, savedexactly). If savevalues! saved value, then saved is true, and if savevalues! saved at the current time point, then savedexactly is true.

The saving priority/order is as follows:

save_on

saveat

force_save

save_everystep

Missing docstring.

Missing docstring for DiffEqBase.savevalues!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.add_tstop!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.add_saveat!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.set_abstol!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.set_reltol!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.reinit!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.auto_dt_reset!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.change_t_via_interpolation!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.addsteps!. Check Documenter's build log for details.

DiffEqBase.reeval_internals_due_to_modification! — Function.

reeval_internals_due_to_modification!(integrator::DDEIntegrator)

Recalculate interpolation data and update ODE integrator after changes by callbacks.

DiffEqBase.set_t! — Function.

set_t!(integrator::DEIntegrator, t::Real)

Set current time point of the integrator to t.

DiffEqBase.set_u! — Function.

set_u!(integrator::DEIntegrator, u)

Set current state of the integrator to u.

DiffEqBase.set_ut! — Function.

set_ut!(integrator::DEIntegrator, u, t)

Set current state of the integrator to u and t

Missing docstring.

Missing docstring for DiffEqBase.addat!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiffEqBase.last_step_failed. Check Documenter's build log for details.

DiffEqBase.check_error — Function.

check_error(integrator)

Check state of integrator and return one of the Return Codes

DiffEqBase.check_error! — Function.

check_error!(integrator)

Same as check_error but also set solution's return code (integrator.sol.retcode) and run postamble!.

Missing docstring.

Missing docstring for DiffEqBase.intervals. Check Documenter's build log for details.

Traits

Missing docstring.

Missing docstring for DiffEqBase.has_reinit. Check Documenter's build log for details.