Extra Functions

FiniteElementDiffEq.getNoise — Function.

getNoise(N,node,elem;noisetype=:White)

Returns a random vector corresponding to the noise type which was chosen.

source

DiffEqBase.numparameters — Function.

numparameters(f)

Returns the number of parameters of f for the method which has the most parameters.

source

DiffEqBase.Tableau — Type.

Tableau: Holds the information for a Runge-Kutta Tableau

source

DiffEqBase.DEProblem — Type.

DEProblem: Defines differential equation problems via its internal functions

source