The Solution Type

Each solver has an appropriate solution type. The solution type holds all of the information about the problem which was solved an its solution. If you enabled save_timeseries=true, then the solver also includes a time-course of the solution captured at every timeseries_steps steps.

The solution type has a lot of built in functionality to help analysis. For example, it has an array interface for accessing the values. We can use

sol[i]

to access the value at timestep i (if the timeseres was saved), and

sol.t[i]

to access the value of t at timestep i. The final value of the simulation, which is always saved, is saved to

sol.u

If the solver allows for dense output (any ODE solver) and dense=true was set for the solving (which is the default), then we can access the approximate value at a time t using the command

sol(t)

If the analytical solution, we also have

sol.u_analytic # final value
sol.timeseries_analytic # timeseries of analytical solution, saved if save_timesseries == true
sol.prob.analytic(t) # The analytic solution at time t

Plotting functionality is provided for each solution type. To plot the solution, simply use

plot(sol)

The plotting function is implemented as a recipe to Plots.jl and as such receives all of the features of a Plots.jl plot.

Solution Types

FiniteElementDiffEq.FEMSolution — Type.

FEMSolution

Holds the data for the solution to a finite element problem.

Fields

fem_mesh::FEMmesh: The finite element mesh the problem was solved on.

u::Array{Float64}: The solution (at the final timepoint)

trueknown::Bool: Boolean flag for if the true solution is given.

u_analytic::AbstractArrayOrVoid: The true solution at the final timepoint.

errors: A dictionary of the error calculations.

appxtrue::Bool: Boolean flag for if u_analytic was an approximation.

timeseries::AbstractArrayOrVoid: u over time. Only saved ifsave_timeseries=true` is specified in the solver.

t::AbstractArrayOrVoid: All the t's in the solution. Only saved if save_timeseries=true is specified in the solver.

prob::DEProblem: Holds the problem object used to define the problem.

save_timeseries::Bool: True if solver saved the extra timepoints.

source

DiffEqBase.DESolution — Type.

PdeSolution: Wrapper for the objects obtained from a solver

source

StochasticDiffEq.SDESolution — Type.

SDESolution

Holds the data for the solution to a SDE problem.

Fields

u::Array{Float64}: The solution (at the final timepoint)

trueknown::Bool: Boolean flag for if the true solution is given.

u_analytic::AbstractArrayOrVoid: The true solution at the final timepoint.

errors: A dictionary of the error calculations.

timeseries::AbstractArrayOrVoid: u over time. Only saved ifsave_timeseries=true` is specified in the solver.

t::AbstractArrayOrVoid: All the t's in the solution. Only saved if save_timeseries=true is specified in the solver.

Ws: All of the W's in the solution. Only saved if save_timeseries=true is specified in the solver.

timeseries_analytic: If save_timeseries=true, saves the solution at each save point.

prob::DEProblem: Holds the problem object used to define the problem.

save_timeseries::Bool: True if solver saved the extra timepoints.

appxtrue::Bool: Boolean flag for if u_analytic was an approximation.

source

OrdinaryDiffEq.ODESolution — Type.

ODESolution

Holds the data for the solution to an ODE problem.

Fields

u::Array{Float64}: The solution (at the final timepoint)

trueknown::Bool: Boolean flag for if the true solution is given.

u_analytic::AbstractArrayOrVoid: The true solution at the final timepoint.

errors: A dictionary of the error calculations.

timeseries::AbstractArrayOrVoid: u over time. Only saved ifsave_timeseries=true` is specified in the solver.

t::AbstractArrayOrVoid: All the t's in the solution. Only saved if save_timeseries=true is specified in the solver.

timeseries_analytic: If save_timeseries=true, saves the solution at each timestep.

prob::DEProblem: Holds the problem object used to define the problem.

save_timeseries::Bool: True if solver saved the extra timepoints.

appxtrue::Bool: Boolean flag for if u_analytic was an approximation.

source

StokesDiffEq.StokesSolution — Type.

StokesSolution

Holds the data for the solution to a Stokes problem.

Fields

u

v

p

u_analytic

vTrue

pTrue

mesh

trueknown

errors

converrors

source

Related Functions

DiffEqDevTools.appxtrue! — Function.

appxtrue!(sol::AbstractODESolution,sol2::TestSolution)

Uses the interpolant from the higher order solution sol2 to approximate errors for sol. If sol2 has no interpolant, only the final error is calculated.

source

appxtrue!(sol::FEMSolution,sol2::FEMSolution)

Adds the solution from sol2 to the FEMSolution object sol. Useful to add a quasi-true solution when none is known by computing once at a very small time/space step and taking that solution as the "true" solution

source

appxtrue!(sol::AbstractODESolution,sol2::AbstractODESolution)

Uses the interpolant from the higher order solution sol2 to approximate errors for sol. If sol2 has no interpolant, only the final error is calculated.

source

FiniteElementDiffEq.FEMSolutionTS — Function.

S = FEMSolutionTS(timeseries::Vector{uType},numvars::Int)S[i][j]` => Variable i at time j.

source