Stokes Solvers

Stokes Equation

DiffEqBase.solve — Method.

solve(prob::StokesProblem,mesh::FDMMesh)

Solves the given stationary Stokes problem on the given finite difference mesh.

Keyword Arguments

converrors: Whether to calculate all of the errors along the convergence. Default is true.

maxiters: Maximum number of iterations before haulting. Default is 100.

alg: The solver algorithm. Default is "dgs". Other option is "multigrid".

level: The number of levels in the Multigrid. Default is 2.

smoothSteps: The number of Gauss-Seidel iterations to do at each smoothing step. Default is 10.

coarseSteps: The number of Gauss-Seidel iterations to do at the coarsegrid. Default is 40.

gsiters: The number of Gauss-Seidel iterations to do at each step. Default is 20.

source